Substitute frame name when needed, use a set for model dictionaries
The frameList's first frame is adopted by the core file, but is legal to occur multiple times. The core generation must substitute with 'clock' in these cases.
This also reveals that the model dictionaries generation could have duplicates, so the iteration is changed to a set rather than a list.